requirements.txt
google-generativeai>=0.5.0
pinecone>=3.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
pyautogui>=0.9.54
pywinauto>=0.6.8
pillow>=10.0.0
opencv-python>=4.8.0
